+++ lib/ui/page/home/page/chat/controller.dart
@@ -43,12 +43,12 @@ import '/api/backend/schema.dart'
 import '/config.dart';
 import '/domain/model/application_settings.dart';
 import '/domain/model/attachment.dart';
-import '/domain/model/chat.dart';
 import '/domain/model/chat_call.dart';
 import '/domain/model/chat_info.dart';
-import '/domain/model/chat_item.dart';
 import '/domain/model/chat_item_quote.dart';
+import '/domain/model/chat_item.dart';
 import '/domain/model/chat_message_input.dart';
+import '/domain/model/chat.dart';
 import '/domain/model/contact.dart';
 import '/domain/model/mute_duration.dart';
 import '/domain/model/precise_date_time/precise_date_time.dart';
@@ -64,7 +64,9 @@ import '/domain/service/call.dart';
 import '/domain/service/chat.dart';
 import '/domain/service/contact.dart';
 import '/domain/service/disposable_service.dart';
+import '/domain/service/my_user.dart';
 import '/domain/service/notification.dart';
+import '/domain/service/session.dart';
 import '/domain/service/user.dart';
 import '/l10n/l10n.dart';
 import '/provider/gql/exceptions.dart'
@@ -112,7 +114,9 @@ class ChatController extends GetxController with IdentityAware {
     this._userService,
     this._settingsRepository,
     this._contactService,
-    this._notificationService, {
+    this._notificationService,
+    this._myUserService,
+    this._sessionService, {
     this.itemId,
     this.onContext,
   });
@@ -324,6 +328,12 @@ class ChatController extends GetxController with IdentityAware {
   /// [AbstractSettingsRepository], used to get the [background] value.
   final AbstractSettingsRepository _settingsRepository;
 
+  /// [MyUserService] to pass to a [MessageFieldController].
+  final MyUserService _myUserService;
+
+  /// [SessionService] to pass to a [MessageFieldController].
+  final SessionService _sessionService;
+
   /// [ContactService] maintaining [ChatContact]s of this [me].
   final ContactService _contactService;
 
@@ -408,8 +418,14 @@ class ChatController extends GetxController with IdentityAware {
   /// Indicates whether a next page of the [elements] is loading.
   RxBool get nextLoading => _fragment?.nextLoading ?? chat!.nextLoading;
 
-  /// Indicates whether the [chat] this [ChatController] is about is a dialog.
+  /// Indicates whether the [chat] this [ChatController] is a dialog.
   bool get isDialog => chat?.chat.value.isDialog == true;
+
+  /// Indicates whether the [chat] this [ChatController] is a monolog.
+  bool get isMonolog => chat?.chat.value.isMonolog == true;
+
+  /// Indicates whether the [chat] this [ChatController] is a support-[Chat].
+  bool get isSupport => chat?.chat.value.isSupport == true;
 
   /// Returns [RxUser] being recipient of this [chat].
   ///
@@ -454,6 +470,10 @@ class ChatController extends GetxController with IdentityAware {
       _chatService,
       _userService,
       _settingsRepository,
+      _authService,
+      _myUserService,
+      _sessionService,
+      _notificationService,
       onChanged: _updateDraft,
       onCall: call,
       onKeyUp: (key) {
@@ -774,6 +794,10 @@ class ChatController extends GetxController with IdentityAware {
         _chatService,
         _userService,
         _settingsRepository,
+        _authService,
+        _myUserService,
+        _sessionService,
+        _notificationService,
         text: item.text?.val,
         onKeyUp: (key) {
           if (key == LogicalKeyboardKey.escape) {
@@ -870,6 +894,7 @@ class ChatController extends GetxController with IdentityAware {
         },
       );
 
+      edit.value?.toggleLogs(isMonolog || isSupport);
       edit.value?.edited.value = item;
       edit.value?.field.focus.requestFocus();
 
@@ -976,6 +1001,8 @@ class ChatController extends GetxController with IdentityAware {
         _chatSubscription = chat!.updates.listen((_) {});
 
         unreadMessages = chat!.chat.value.unreadCount;
+
+        send.toggleLogs(isMonolog || isSupport);
 
         await chat!.ensureDraft();
         final ChatMessage? draft = chat!.draft.value;