+++ lib/ui/page/home/page/chat/forward/controller.dart
@@ -1,5 +1,7 @@
 // Copyright © 2022-2026 IT ENGINEERING MANAGEMENT INC,
 //                       <https://github.com/team113>
+// Copyright © 2025 Ideas Networks Solutions S.A.,
+//                       <https://github.com/tapopa>
 //
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the GNU Affero General Public License v3.0 as published by the
@@ -106,6 +108,10 @@ class ChatForwardController extends GetxController {
       _chatService,
       _userService,
       _settingsRepository,
+      null,
+      null,
+      null,
+      null,
       text: text,
       attachments: attachments,
       canPin: false,