+++ lib/ui/page/home/page/chat/message_field/view.dart
@@ -1,5 +1,7 @@
 // Copyright © 2022-2026 IT ENGINEERING MANAGEMENT INC,
 //                       <https://github.com/team113>
+// Copyright © 2025 Ideas Networks Solutions S.A.,
+//                       <https://github.com/tapopa>
 //
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the GNU Affero General Public License v3.0 as published by the
@@ -150,7 +152,15 @@ class MessageFieldView extends StatelessWidget {
     return GetBuilder(
       init:
           controller ??
-          MessageFieldController(Get.find(), Get.find(), Get.find()),
+          MessageFieldController(
+            Get.find(),
+            Get.find(),
+            Get.find(),
+            Get.find(),
+            Get.find(),
+            Get.find(),
+            Get.find(),
+          ),
       global: false,
       builder: (MessageFieldController c) {
         return CallbackShortcuts(