+++ lib/ui/page/style/page/widgets/section/fields.dart
@@ -1,5 +1,7 @@
 // Copyright © 2022-2026 IT ENGINEERING MANAGEMENT INC,
 //                       <https://github.com/team113>
+// Copyright © 2025 Ideas Networks Solutions S.A.,
+//                       <https://github.com/tapopa>
 //
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the GNU Affero General Public License v3.0 as published by the
@@ -124,7 +126,15 @@ class FieldsSection {
       ),
       Headline(
         child: MessageFieldView(
-          controller: MessageFieldController(null, null, null),
+          controller: MessageFieldController(
+            null,
+            null,
+            null,
+            null,
+            null,
+            null,
+            null,
+          ),
         ),
       ),
       Headline(