+++ test/widget/chat_edit_message_test.dart
@@ -33,6 +33,8 @@ import 'package:messenger/domain/model/session.dart';
 import 'package:messenger/domain/model/user.dart';
 import 'package:messenger/domain/repository/auth.dart';
 import 'package:messenger/domain/repository/chat.dart';
+import 'package:messenger/domain/repository/my_user.dart';
+import 'package:messenger/domain/repository/session.dart';
 import 'package:messenger/domain/repository/settings.dart';
 import 'package:messenger/domain/service/auth.dart';
 import 'package:messenger/domain/service/call.dart';
@@ -40,6 +42,7 @@ import 'package:messenger/domain/service/chat.dart';
 import 'package:messenger/domain/service/contact.dart';
 import 'package:messenger/domain/service/my_user.dart';
 import 'package:messenger/domain/service/notification.dart';
+import 'package:messenger/domain/service/session.dart';
 import 'package:messenger/domain/service/user.dart';
 import 'package:messenger/provider/drift/account.dart';
 import 'package:messenger/provider/drift/background.dart';
@@ -53,10 +56,12 @@ import 'package:messenger/provider/drift/chat_member.dart';
 import 'package:messenger/provider/drift/credentials.dart';
 import 'package:messenger/provider/drift/draft.dart';
 import 'package:messenger/provider/drift/drift.dart';
+import 'package:messenger/provider/drift/geolocation.dart';
 import 'package:messenger/provider/drift/locks.dart';
 import 'package:messenger/provider/drift/monolog.dart';
 import 'package:messenger/provider/drift/my_user.dart';
 import 'package:messenger/provider/drift/secret.dart';
+import 'package:messenger/provider/drift/session.dart';
 import 'package:messenger/provider/drift/settings.dart';
 import 'package:messenger/provider/drift/slugs.dart';
 import 'package:messenger/provider/drift/user.dart';
@@ -69,6 +74,7 @@ import 'package:messenger/store/call.dart';
 import 'package:messenger/store/chat.dart';
 import 'package:messenger/store/contact.dart';
 import 'package:messenger/store/my_user.dart';
+import 'package:messenger/store/session.dart';
 import 'package:messenger/store/settings.dart';
 import 'package:messenger/store/user.dart';
 import 'package:messenger/themes.dart';
@@ -79,6 +85,7 @@ import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 
 import '../mock/audio_utils.dart';
+import '../mock/geo_provider.dart';
 import '../mock/platform_utils.dart';
 import 'chat_edit_message_test.mocks.dart';
 import 'extension/rich_text.dart';
@@ -98,6 +105,7 @@ void main() async {
   Config.disableDragArea = true;
 
   var graphQlProvider = MockGraphQlProvider();
+  when(graphQlProvider.connected).thenReturn(RxBool(true));
   when(graphQlProvider.disconnect()).thenAnswer((_) => () {});
   when(
     graphQlProvider.onStart,
@@ -368,6 +376,8 @@ void main() async {
   final locksProvider = Get.put(LockDriftProvider(common));
   final secretsProvider = Get.put(RefreshSecretDriftProvider(common));
   final slugProvider = Get.put(SlugDriftProvider(common));
+  final sessionProvider = Get.put(SessionDriftProvider(common, scoped));
+  final geoProvider = Get.put(GeoLocationDriftProvider(common));
 
   Widget createWidgetForTesting({required Widget child}) {
     return MaterialApp(
@@ -462,15 +472,32 @@ void main() async {
     );
     Get.put(ContactService(contactRepository));
 
-    MyUserRepository myUserRepository = MyUserRepository(
-      graphQlProvider,
-      myUserProvider,
-      blocklistRepository,
-      userRepository,
-      accountProvider,
-      me: const UserId('me'),
-    );
+    final AbstractMyUserRepository myUserRepository =
+        Get.put<AbstractMyUserRepository>(
+          MyUserRepository(
+            graphQlProvider,
+            myUserProvider,
+            blocklistRepository,
+            userRepository,
+            accountProvider,
+            me: const UserId('me'),
+          ),
+        );
     Get.put(MyUserService(authService, myUserRepository));
+
+    final AbstractSessionRepository sessionRepository =
+        Get.put<AbstractSessionRepository>(
+          SessionRepository(
+            graphQlProvider,
+            accountProvider,
+            versionProvider,
+            sessionProvider,
+            geoProvider,
+            MockedGeoLocationProvider(),
+            me: const UserId('me'),
+          ),
+        );
+    Get.put(SessionService(sessionRepository));
 
     Get.put(UserService(userRepository));
     ChatService chatService = Get.put(ChatService(chatRepository, authService));